problem5/documentblockchain/x/documentblockchain/types/messages_document.go
package types

import (
	errorsmod "cosmossdk.io/errors"
	sdk "github.com/cosmos/cosmos-sdk/types"
	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
)

var _ sdk.Msg = &MsgCreateDocument{}

func NewMsgCreateDocument(creator string, title string, body string) *MsgCreateDocument {
	return &MsgCreateDocument{
		Creator: creator,
		Title:   title,
		Body:    body,
	}
}

func (msg *MsgCreateDocument) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}

var _ sdk.Msg = &MsgUpdateDocument{}

func NewMsgUpdateDocument(creator string, id uint64, title string, body string) *MsgUpdateDocument {
	return &MsgUpdateDocument{
		Id:      id,
		Creator: creator,
		Title:   title,
		Body:    body,
	}
}

func (msg *MsgUpdateDocument) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}

var _ sdk.Msg = &MsgDeleteDocument{}

func NewMsgDeleteDocument(creator string, id uint64) *MsgDeleteDocument {
	return &MsgDeleteDocument{
		Id:      id,
		Creator: creator,
	}
}

func (msg *MsgDeleteDocument) ValidateBasic() error {
	_, err := sdk.AccAddressFromBech32(msg.Creator)
	if err != nil {
		return errorsmod.Wrapf(sdkerrors.ErrInvalidAddress, "invalid creator address (%s)", err)
	}
	return nil
}
